主页模糊分页查询待完成，已完成用户模糊分页查询

+++ src/views/PersonalCenter/PinBoards/Pinboards.vue
@@ -1,20 +1,25 @@
+<!--
+[用一个div画出关闭图标](https://segmentfault.com/a/1190000008955548)
+-->
 <template>
   <el-container>
     <WaterfallMain
         :images="images"
-        :showCardFooter="true"
         :showAvatar="false"
+        :showSaveBtnInWaterfall="false"
         :handleDelete="handleDelete"
         :handle-click="handleClick"
         :handleEdit="handleEdit"/>
     <Dialog
         :showDialog.sync="showDialog"
-        :commentsOrDreamForm="true"
-        :showSelect="true"/>
-    <!--[admin vue页面右下角添加客服按钮](https://blog.csdn.net/thc1987/article/details/106623974)-->
+        :showCommentsOrDreamForm="true"
+        />
+    <!--
+      [admin vue页面右下角添加客服按钮](https://blog.csdn.net/thc1987/article/details/106623974)
+    -->
     <div id="add_pins_button" v-show="true">
       <el-popover placement="left-end" trigger="click">
-        <el-button type="text" style="text-align: center;" @click="showDialog = !showDialog">Add Your Pins</el-button>
+        <el-button type="text" style="text-align: center;" @click="initDreamForm">Add Your Pins</el-button>
         <label slot="reference" class="top_btn" title="Add Your Pins"></label>
       </el-popover>
     </div>
@@ -26,67 +31,70 @@ import Header from "../../../components/Header/Header";
 import WaterfallMain from "../../../components/WaterfallMain/WaterfallMain";
 import Dialog from "../../../components/Dialog/Dialog";
 import {EventBus} from "../../../apis/eventBus"
-import {getRequest,deleteRequest} from "../../../apis/api";
+import {getRequest, deleteRequest} from "../../../apis/api";
 
 export default {
   components: {Header, WaterfallMain, Dialog},
   mounted() {
     // 这里的 images 初始化从数据库获取 TODO
-    getRequest('get-pinboards').then(res=>{
-      this.images = res.object
-      console.log(this.images)
-    }).catch(err=>{
+    getRequest('get-pinboards').then(res => {
+      if (res) {
+        this.images = res.object
+      } else {
+        this.$message('没有查询到您的愿望卡！')
+      }
+    }).catch(err => {
       console.log(err)
-    })
-    EventBus.$on('getDreamForm',data => {
-      console.log(data)
-      this.images.push(data)
-      console.log(this.images.length)
-    })
-    EventBus.$on('getShowDialog',showDialog => {
-      this.showDialog = showDialog
-      console.log(this.showDialog)
-      })
+    });
+    EventBus.$on('getShowDialog', value => {
+      this.showDialog = value
+      console.log(showDialog)
+    });
   },
   data() {
     return {
-      // dreamForm: {
-      //   pinboardSharable: true,
-      //   pinboardBgimgUrl: '',
-      //   pinboardTitle: '',
-      //   pinboardContent: '',
-      //   pinUsername: '',
-      //   pinUserAvatarUrl: '',
-      // },
-      userInfo: JSON.parse(window.sessionStorage.getItem('userInfo')),
+      dreamForm: {},
+      userInfo: this.$store.state.userInfo,
       search: '',
       list: [],
       showDialog: false,
-      imageUrl: "https://i.pinimg.com/236x/4d/ba/24/4dba24872bed032eeaf85e51bbd502b9.jpg",
+      imageUrl: '',
       images: [],
     }
   },
   methods: {
+    initDreamForm() {
+      this.showDialog = !this.showDialog
+      EventBus.$emit('initImageUrl', '')
+      EventBus.$emit('initDreamForm', {})
+      EventBus.$emit('saveOrUpdate',true)
+    },
     handleClick(item) {
       this.$message.warning(item.pinboardBgimgUrl);
     },
     // * 编辑
     handleEdit(item) {
-      // 弹出编辑框，TODO 发送请求向数据库更新一个 pinboard
-      EventBus.$emit('getShowDialog',true)
-      EventBus.$emit('initImageUrl',item.pinboardBgimgUrl)
-      this.$message.success('编辑');
+      // TODO 弹出编辑框 发送请求向数据库更新一个 pinboard
+      this.showDialog = true
+      EventBus.$emit('initImageUrl', item.pinboardBgimgUrl)
+      EventBus.$emit('initDreamFormFromPinboards', item)
+      EventBus.$emit('saveOrUpdate', false)
     },
     // * 删除
     handleDelete(item) {
-      // TODO 发送请求向数据库移除一个 pinboard 同时移除其下的评论和点赞信息
-      deleteRequest(`delete-one-pinboard/${item.pinboardId}`).then(res=>{
-        console.log(res)
-      }).catch(err=>{
+      this.$confirm('确认删除吗？', '删除一个pin', 'warning').then(res => {
+        if (res) {
+          // TODO 发送请求向数据库移除一个 pinboard 同时移除其下的评论和点赞信息
+          deleteRequest(`delete-one-pinboard/${item.pinboardId}`).then(res => {
+            console.log(res)
+          }).catch(err => {
+            console.log(err)
+          })
+          console.log('Pinboards->handleDelete()',item.pinboardId)
+        }
+      }).catch(err => {
         console.log(err)
       })
-      console.log(item.pinboardId)
-      this.$message.error('删除');
     },
   },
   // [vue中event bus被触发多次问题](https://segmentfault.com/q/1010000009710635)
@@ -105,6 +113,7 @@ export default {
   z-index: 999;
   width: 60px;
   height: 60px;
+
   .top_btn {
     width: 60px;
     height: 60px;
